Do not drop + create indexes on every insert (BRAINSTORM-1162)
Dropping the index on starup may be beneficial if the cache is not
even close to being up to date. However, doing the same when the cache
is close to being up to date is not good. Disabled the current drop+create
code. Should add code later on to determine whether

  1) the update is the first one
  2) the update is so big that the index should be dropped and rebuilt

Note that the engine will not register itself ready until the cache
has been updated at least once, hence temporatily dropping the indices
during a heavy init is safe.