Keep track of rows inserted into the observation cache in one place only
Previously the insert cache objects were in the SpatiaLite
object. However, insertion requests a new SpatiaLite object from the
SpatiaLiteCache connection pool instead of reusing the same
connection.  Hence the proper place for the cache objects is in
SpatiaLiteCache itself, it will just pass the object to the SpatiaLite
instance to be updated.