Fixed sorting of nearby stations (BRAINSTORM-1367)
NearTree randomizes the data structure, and hence its output may
change if stations are at the exact same distance. This happens for
example for road weather stations, whose coordinates are identical.

The code now uses lexocographic sort for the nearby stations, the station
id is compared if the distances are equal.